+++ media/filters/file_data_source.h
@@ -26,7 +26,7 @@ class MEDIA_EXPORT FileDataSource : public DataSource {
   // Implementation of DataSource.
   virtual void set_host(DataSourceHost* host) OVERRIDE;
   virtual void Stop(const base::Closure& callback) OVERRIDE;
-  virtual void Read(int64 position, size_t size, uint8* data,
+  virtual void Read(int64 position, int size, uint8* data,
                     const DataSource::ReadCB& read_cb) OVERRIDE;
   virtual bool GetSize(int64* size_out) OVERRIDE;
   virtual bool IsStreaming() OVERRIDE;
